gave up on cubic splines, going with makima interpolation instead
